Make it a gem! :gem:
Shuffle things around a bit and add a gemspec and related stuff.
