+++ bin/deploy-report
@@ -13,20 +13,16 @@
 #
 # Pull up the help message to learn about the usage of this script:
 #
-#         ./deploy-report.rb --help
+#         ./bin/deploy-report --help
 #
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 
-require 'uri'
-require 'net/http'
-require 'json'
+require 'appsignal_report'
 require 'optparse'
-
-require_relative 'lib/appsignal_deploy_report'
 
 options = { format: :text, app_id: nil }
 parser = OptionParser.new do |parser|
-  parser.banner = 'Usage: APPSIGNAL_API_TOKEN=XXX ./deploy-report.rb [options]'
+  parser.banner = 'Usage: APPSIGNAL_API_TOKEN=XXX ./bin/deploy-report [options]'
   parser.separator ''
   parser.separator 'Specific options:'
   parser.on('-i ID',
@@ -50,7 +46,7 @@ parser = OptionParser.new do |parser|
 end
 parser.parse!
 
-report = AppsignalDeployReport.new(
+report = AppsignalReport::DeployReport.new(
   api_token: ENV['APPSIGNAL_API_TOKEN'],
   app_id: options[:app_id]
 )
