+++ lib/appsignal_report/version.rb
@@ -1,0 +1,3 @@
+module AppsignalReport
+  VERSION = '0.1.0'
+end
